Crate unicode_truncate

Unicode-aware algorithm to pad or truncate str in terms of displayed width.

See the UnicodeTruncateStr trait for new methods available on str.

Examples

Safely truncate string to display width even not at character boundaries.

use unicode_truncate::UnicodeTruncateStr;
assert_eq!("你好吗".unicode_truncate(5), ("你好", 4));

Making sure the string is displayed in exactly number of columns by combining padding and truncating.

use unicode_truncate::UnicodeTruncateStr;
use unicode_truncate::Alignment;
use unicode_width::UnicodeWidthStr;

let str = "你好吗".unicode_pad(5, Alignment::Left, true);
assert_eq!(str, "你好 ");
assert_eq!(str.width(), 5);

Enums

Alignment

Defines the alignment for truncation and padding.

Traits

UnicodeTruncateStr

Methods for padding or truncating using displayed width of Unicode strings.